+++ R/uniset_functions.R
@@ -68,6 +68,13 @@ test_getLocalStn <- function() {
 #' to the desired folder can be selected interactively. 
 #' @param path Character length one. The path to the folder where the settings-
 #' home folder should be located. Defaults to \code{NULL}.
+#' @param path Character length one, holding the path to the location where the 
+#' folder containing the settings.R file should be located. Defaults to 
+#' \code{NULL}. The path has to be given either in forward-slashes 
+#' (\code{/Users/Name/aFolder} on Unix machines and \code{C:/Users/Name/aFolder} 
+#' on windows machines), or in double back-slashes 
+#' \code{C:\\\\Users\\\\Name\\\\aFolder} on windows machines.
+#' If left at the default \code{NULL}, the location should be selectable interactively.
 #' @return Called for its side-effects, i.e. performing the required setup 
 #' regarding the settings.R file. 
 #' @family Helper Functions
@@ -135,6 +142,15 @@ autoUpS <- function(cfs=getstn()$defCfs) {
 	stn <- uniset::uniset_autoUpS(get("uniset_handover"),
 		setupFunc="ap2_settings_setup")
 	return(invisible(stn))
+} # EOF
+
+
+#' @title Test for Uniset
+#' @description Is testing if the uniset integration is set up correctly.
+#' @return A list of length 4.
+#  #' @export
+testUniset <- function() {
+	uniset::uniset_test(uniset_handover)
 } # EOF
 
 #######################